#26 . Test de borrar idea.
