+++ package.json
@@ -29,6 +29,7 @@
   "devDependencies": {
     "angular-mocks": "^1.5.5",
     "chai": "^3.5.0",
+    "chai-as-promised": "^5.3.0",
     "express-mocks-http": "0.0.11",
     "karma": "^0.13.22",
     "karma-chai": "^0.1.0",
@@ -40,12 +41,12 @@
     "karma-sinon": "^1.0.5",
     "karma-sinon-chai": "^1.2.0",
     "mocha": "^2.4.5",
+    "mockgoose": "^6.0.2",
     "node-mocks-http": "^1.5.2",
     "phantomjs-prebuilt": "^2.1.7",
-    "mockgoose": "^6.0.2",
     "proxyquire": "^1.7.7",
-    "supertest": "^1.2.0",
     "sinon": "^1.17.4",
-    "sinon-chai": "^2.8.0"
+    "sinon-chai": "^2.8.0",
+    "supertest": "^1.2.0"
   }
 }
